block: elevator: Initialize blk-mq devices at compile time
IS_ENABLED() is a runtime macro and ifdef/if defined() is a compile
time macro which is more preferable than the other as it will
ignore the codes which won't satisfy the condition directly at
compile time instead of checking every single condition at runtime
like IS_ENABLED().

This is a small miscellaneous improvement.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: pix106 <sbordenave@gmail.com>