lib/crypto/Kconfig: Remove unneeded selection of XOR_BLOCKS
CRYPTO_LIB_CHACHA_GENERIC doesn't need to select XOR_BLOCKS.  It perhaps
was thought that it's needed for __crypto_xor, but that's not the case.

Enabling XOR_BLOCKS is problematic because the XOR_BLOCKS code runs a
benchmark when it is initialized.  That causes a boot time regression on
systems that didn't have it enabled before.

Therefore, remove this unnecessary and problematic selection.

Bug: 242983064
Fixes: c3632866fc7f ("lib/crypto: add prompts back to crypto libraries")
Test: asit/perf/boottime_test(BootAnimationShownTiming-mean)
Cc: stable@vger.kernel.org
Signed-off-by: Eric Biggers <ebiggers@google.com>
Signed-off-by: Lucas Wei <lucaswei@google.com>
Change-Id: I32938d078243176d218ebef467f4a0045ea7bd8c
Signed-off-by: pix106 <sbordenave@gmail.com>