qcacld-3.0: Avoid sending RSNXE to legacy APs
Userspace may send RSNXE also in connect request irrespective
of the connecting AP capabilities. This allows the driver to
choose best candidate based on score. But the chosen candidate
may not support the RSNXE features. Station is not supposed to
include the RSNX IE in assoc request in such cases as legacy APs
may misbehave due to the new IE. It's observed that few legacy
APs which don't support the RSNXE reject the connection at EAPOL
stage. So, strip the IE when the connection is legacy than WPA3
mode.

Change-Id: Id6e8095aa73367e69f75dd54a67cdd4dec1879fb
CRs-Fixed: 3828317